+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/workshop/AddNewWorkshop.kt
@@ -275,7 +275,7 @@ fun AddNewWorkshopScreen(
                         modifier =
                             Modifier
                                 .size(width = 112.dp, height = 40.dp),
-                        onClick = { navController.navigate(Screen.WorkshopsList.route) },
+                        onClick = {navController.popBackStack() },
                     )
 
                     Spacer(modifier = Modifier.width(16.dp))