+++ CHANGELOG.md
@@ -3,7 +3,15 @@
 ### Added
 - US for modify User with MVVM and clean architecture principles.
 
+### Fixed
+- Fixed search bar in workshop list.
+- Error now properly appearing when registering new beneficiary
+
 ## [0.1.0] - 18/11/2025
+
+### Added
+- ProductList screen to navgraph
+- Modify product implementation
 
 ### Fixed
 - Fixed the possibility to add or modify one supply with the same name